license_portal/requirements.txt
Django==3.2
psycopg2
django-cors-headers==3.14.0
python-dateutil==2.8.2
djangorestframework==3.14.0